Modifies Player Script/ small alteration to animator conditions
Edit's code to address movement of player exclusively, allowing the
player walk left and right using input other than just the left and
right arrow key.

Also adds the condition to the animator for the fall animation to
transition to the land animation when the "Landing" bool variable
becomes true.
(This does not solve the glitch in the land animation activating early)